final updated models and moved to different files
